bug fixes in imports, start docs
