+++ docs/content/earthquake.rst
@@ -1,0 +1,8 @@
+.. _earthquake:
+
+Earthquake
+==========
+
+.. autoclass:: geoana.earthquake
+    :members:
+    :undoc-members:
