+++ docs/content/em.rst
@@ -1,0 +1,8 @@
+.. _em:
+
+Electromagnetics
+================
+
+.. autoclass:: geoana.em
+    :members:
+    :undoc-members:
