+++ docs/index.rst
@@ -7,6 +7,8 @@ Contents:
 .. toctree::
    :maxdepth: 2
 
+   content/earthquake
+   content/em
 
 
 Indices and tables
