+++ geoana/em/fdem.py
@@ -6,6 +6,7 @@ from __future__ import unicode_literals
 from scipy.constants import mu_0, pi, epsilon_0
 import numpy as np
 import warnings
+import properties
 
 from .base import BaseElectricDipole, BaseMagneticDipole, BaseEM
 
@@ -161,6 +162,7 @@ class BaseFDEMDipoleWholeSpace(BaseFDEM):
     """
     Base FDEM Dipole
     """
+    pass
 
 
 class ElectricDipoleWholeSpace(
@@ -250,7 +252,7 @@ class ElectricDipoleWholeSpace(
         return self.mu * self.magnetic_field(xyz)
 
 
-class MagneticDipoleWholeSpace(BaseMagneticDipole, BaseFDEM)
+class MagneticDipoleWholeSpace(BaseMagneticDipole, BaseFDEM):
     """
     Harmonic magnetic dipole in a whole space.
     """
@@ -266,58 +268,6 @@ class MagneticDipoleWholeSpace(BaseMagneticDipole, BaseFDEM)
     def magnetic_flux_density(self, xyz):
         pass
 
-
-def E_from_EDWS(XYZ, srcLoc, sig, f, current=1., length=1., orientation='X', kappa=0., epsr=1., t=0.):
-    """E_from_EDWS
-    Computing the analytic electric fields (E) from an electrical dipole in a wholespace
-    - You have the option of computing E for multiple frequencies at a single reciever location
-      or a single frequency at multiple locations
-
-    :param numpy.array XYZ: reciever locations at which to evaluate E
-    :param float epsr: relative permitivitty value (unitless),  default is 1.0
-    :rtype: numpy.array
-    :return: Ex, Ey, Ez: arrays containing all 3 components of E evaluated at the specified locations and frequencies.
-    """
-
-    mu = mu_0*(1+kappa)
-    epsilon = epsilon_0*epsr
-    sig_hat = sig + 1j*omega(f)*epsilon
-
-    XYZ = Utils.asArray_N_x_Dim(XYZ, 3)
-    # Check
-    if XYZ.shape[0] > 1 & f.shape[0] > 1:
-        raise Exception('I/O type error: For multiple field locations only a single frequency can be specified.')
-
-    dx = XYZ[:, 0] - srcLoc[0]
-    dy = XYZ[:, 1] - srcLoc[1]
-    dz = XYZ[:, 2] - srcLoc[2]
-
-    r = np.sqrt(dx**2. + dy**2. + dz**2.)
-    # k  = np.sqrt( -1j*2.*pi*f*mu*sig )
-    k = np.sqrt(omega(f)**2. * mu * epsilon - 1j * omega(f) * mu * sig)
-
-    front = current * length / (4.*pi*sig_hat* r**3) * np.exp(-1j*k*r)
-    mid   = -k**2 * r**2 + 3*1j*k*r + 3
-
-    if orientation.upper() == 'X':
-        Ex = front*((dx**2 / r**2)*mid + (k**2 * r**2 -1j*k*r-1.))
-        Ey = front*(dx*dy  / r**2)*mid
-        Ez = front*(dx*dz  / r**2)*mid
-        return Ex, Ey, Ez
-
-    elif orientation.upper() == 'Y':
-        #  x--> y, y--> z, z-->x
-        Ey = front*((dy**2 / r**2)*mid + (k**2 * r**2 -1j*k*r-1.))
-        Ez = front*(dy*dz  / r**2)*mid
-        Ex = front*(dy*dx  / r**2)*mid
-        return Ex, Ey, Ez
-
-    elif orientation.upper() == 'Z':
-        # x --> z, y --> x, z --> y
-        Ez = front*((dz**2 / r**2)*mid + (k**2 * r**2 -1j*k*r-1.))
-        Ex = front*(dz*dx  / r**2)*mid
-        Ey = front*(dz*dy  / r**2)*mid
-        return Ex, Ey, Ez
 
 
 def MagneticDipoleFields(srcLoc, obsLoc, component, orientation='Z', moment=1., mu=mu_0):
